added stored relation
